requirements.txt
django >=1.10,<1.11.99
django-app-settings
idna
fred-pylogger
fred-pyfco
six
